+++ packages/auth/src/api/index.test.ts
@@ -95,18 +95,26 @@ describe('api/_performApiRequest', () => {
     });
 
     it('should include whatever headers the auth impl attaches', async () => {
-      sinon.stub(auth, '_getAdditionalHeaders').returns(Promise.resolve({
-        'look-at-me-im-a-header': 'header-value',
-        'anotherheader': 'header-value-2',
-      }));
+      sinon.stub(auth, '_getAdditionalHeaders').returns(
+        Promise.resolve({
+          'look-at-me-im-a-header': 'header-value',
+          'anotherheader': 'header-value-2'
+        })
+      );
 
       const mock = mockEndpoint(Endpoint.SIGN_UP, serverResponse);
-      await _performApiRequest<
-        typeof request,
-        typeof serverResponse
-      >(auth, HttpMethod.POST, Endpoint.SIGN_UP, request);
-      expect(mock.calls[0].headers.get('look-at-me-im-a-header')).to.eq('header-value');
-      expect(mock.calls[0].headers.get('anotherheader')).to.eq('header-value-2');
+      await _performApiRequest<typeof request, typeof serverResponse>(
+        auth,
+        HttpMethod.POST,
+        Endpoint.SIGN_UP,
+        request
+      );
+      expect(mock.calls[0].headers.get('look-at-me-im-a-header')).to.eq(
+        'header-value'
+      );
+      expect(mock.calls[0].headers.get('anotherheader')).to.eq(
+        'header-value-2'
+      );
     });
 
     it('should set the framework in clientVersion if logged', async () => {
@@ -235,8 +243,7 @@ describe('api/_performApiRequest', () => {
         Endpoint.SIGN_UP,
         request
       );
-      await expect(promise)
-        .to.be.rejectedWith(FirebaseError, 'Text text text');
+      await expect(promise).to.be.rejectedWith(FirebaseError, 'Text text text');
     });
 
     it('should handle unknown server errors', async () => {
@@ -316,7 +323,10 @@ describe('api/_performApiRequest', () => {
         request
       );
       clock.tick(DEFAULT_API_TIMEOUT_MS.get() + 1);
-      await expect(promise).to.be.rejectedWith(FirebaseError, 'auth/network-request-failed');
+      await expect(promise).to.be.rejectedWith(
+        FirebaseError,
+        'auth/network-request-failed'
+      );
       clock.restore();
     });
 
@@ -499,9 +509,7 @@ describe('api/_performApiRequest', () => {
 
     it('does not attach the tenant ID at all if not specified', () => {
       auth.tenantId = null;
-      expect(
-        _addTidIfNecessary<Record<string, string>>(auth, { foo: 'bar' })
-      )
+      expect(_addTidIfNecessary<Record<string, string>>(auth, { foo: 'bar' }))
         .to.eql({
           foo: 'bar'
         })