+++ packages/database-compat/src/api/Reference.ts
@@ -164,9 +164,7 @@ export class DataSnapshot implements Compat<ModularDataSnapshot> {
    * @returns True if forEach was canceled by action returning true for
    * one of the child nodes.
    */
-  forEach(
-    action: (snapshot: DataSnapshot & { key: string }) => boolean | void
-  ): boolean {
+  forEach(action: (snapshot: DataSnapshot) => boolean | void): boolean {
     validateArgCount('DataSnapshot.forEach', 1, 1, arguments.length);
     validateCallback('DataSnapshot.forEach', 'action', action, false);
     return this._delegate.forEach(expDataSnapshot =>