+++ packages/firestore/test/integration/api/validation.test.ts
@@ -51,7 +51,8 @@ import {
   updateDoc,
   where,
   or,
-  and
+  and,
+  newTestApp
 } from '../util/firebase_export';
 import {
   apiDescribe,
@@ -158,19 +159,19 @@ apiDescribe('Validation:', (persistence: boolean) => {
 
     validationIt(persistence, 'enforces minimum cache size', () => {
       expect(() =>
-        newTestFirestore('test-project', undefined, { cacheSizeBytes: 1 })
+        newTestFirestore(newTestApp('test-project'), { cacheSizeBytes: 1 })
       ).to.throw('cacheSizeBytes must be at least 1048576');
     });
 
     validationIt(persistence, 'garbage collection can be disabled', () => {
       // Verify that this doesn't throw.
-      newTestFirestore('test-project', undefined, {
+      newTestFirestore(newTestApp('test-project'), {
         cacheSizeBytes: /* CACHE_SIZE_UNLIMITED= */ -1
       });
     });
 
     validationIt(persistence, 'useEmulator can set host and port', () => {
-      const db = newTestFirestore('test-project');
+      const db = newTestFirestore(newTestApp('test-project'));
       // Verify that this doesn't throw.
       connectFirestoreEmulator(db, 'localhost', 9000);
     });
@@ -193,7 +194,7 @@ apiDescribe('Validation:', (persistence: boolean) => {
       persistence,
       'useEmulator can set mockUserToken object',
       () => {
-        const db = newTestFirestore('test-project');
+        const db = newTestFirestore(newTestApp('test-project'));
         // Verify that this doesn't throw.
         connectFirestoreEmulator(db, 'localhost', 9000, {
           mockUserToken: { sub: 'foo' }
@@ -205,7 +206,7 @@ apiDescribe('Validation:', (persistence: boolean) => {
       persistence,
       'useEmulator can set mockUserToken string',
       () => {
-        const db = newTestFirestore('test-project');
+        const db = newTestFirestore(newTestApp('test-project'));
         // Verify that this doesn't throw.
         connectFirestoreEmulator(db, 'localhost', 9000, {
           mockUserToken: 'my-mock-user-token'